requirements.txt
sqlalchemy
openai
python-dotenv
pydantic
fastapi
typing_extensions
requests
uvicorn
